--- config.py
@@ -1,3 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)  # Initialising a flask application
